bridges/praetor-execd/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "praetor-execd"
version = "0.1.0"
description = "Host-side executor bridge for Praetor subscription executor mode"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "PyYAML>=6.0.0,<7.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "httpx>=0.27.0,<1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["praetor_execd*"]
